search: Add an example of searches, vizualization, Dashboard
The current instance of ElasticSearch comes without examples. It can be
quiet hard to comprehend how it works at first to create vizualisations
and Dashboards.

Add a json file exported from KernelCI's ElasticSearch. This json file
contains searches, visualization and an usefull Dashboard.

The idea is to get people using the same Dashboard. And hopefully even
do pull requests on it do improve it.

Signed-off-by: Loys Ollivier <lollivier@baylibre.com>